+++ src/main/java/com/clover/habbittracker/global/security/jwt/JwtFilter.java
@@ -1,4 +1,4 @@
-package com.clover.habbittracker.global.jwt;
+package com.clover.habbittracker.global.security.jwt;
 
 import java.io.IOException;
 import java.util.List;
