+++ src/main/java/com/clover/habbittracker/global/security/oauth/OauthService.java
@@ -1,10 +1,10 @@
-package com.clover.habbittracker.global.oauth;
+package com.clover.habbittracker.global.security.oauth;
 
 import org.springframework.stereotype.Service;
 
-import com.clover.habbittracker.global.jwt.JwtProvider;
+import com.clover.habbittracker.global.security.jwt.JwtProvider;
 import com.clover.habbittracker.domain.member.service.MemberService;
-import com.clover.habbittracker.global.oauth.dto.SocialUser;
+import com.clover.habbittracker.global.security.oauth.dto.SocialUser;
 
 import lombok.RequiredArgsConstructor;
