+++ src/main/java/com/clover/habbittracker/global/security/oauth/dto/GoogleUser.java
@@ -1,4 +1,4 @@
-package com.clover.habbittracker.global.oauth.dto;
+package com.clover.habbittracker.global.security.oauth.dto;
 
 import java.util.Map;
